Unpack the asyncio helper functions into a separate module
